ACMS-858: Remove content test from acquia_cms_site_studio module
