+++ modules/acquia_cms_search/tests/src/ExistingSite/ArticleListTest.php
@@ -1,47 +1,47 @@
 <?php
 
-namespace Drupal\Tests\acquia_cms_site_studio\ExistingSite;
+namespace Drupal\Tests\acquia_cms_search\ExistingSite;
 
 use Drupal\Core\Entity\Query\QueryInterface;
 use Drupal\Tests\acquia_cms_common\ExistingSite\ContentTypeListTestBase;
 use Drupal\views\Entity\View;
 
 /**
- * Tests the "all events" listing page.
+ * Tests the "all articles" listing page.
  *
- * @group acquia_cms
  * @group acquia_cms_site_studio
+ * @group acquia_cms
  * @group low_risk
  * @group pr
  * @group push
  */
-class EventListTest extends ContentTypeListTestBase {
+class ArticleListTest extends ContentTypeListTestBase {
 
   /**
    * {@inheritdoc}
    */
-  protected $nodeType = 'event';
+  protected $nodeType = 'article';
 
   /**
    * {@inheritdoc}
    */
   protected function getView() : View {
-    return View::load('events');
+    return View::load('articles');
   }
 
   /**
    * {@inheritdoc}
    */
   protected function visitListPage($langcode = NULL) : void {
-    $page = $langcode ? "/$langcode/events" : "/events";
+    $page = $langcode ? "/$langcode/articles" : "/articles";
     $this->drupalGet($page);
   }
 
   /**
    * {@inheritdoc}
    */
-  protected function getQuery() : QueryInterface {
-    return parent::getQuery()->sort('field_event_start')->sort('title');
+  protected function getQuery(): QueryInterface {
+    return parent::getQuery()->sort('created', 'DESC');
   }
 
 }
