+++ modules/acquia_cms_search/tests/src/ExistingSite/EventListTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Drupal\Tests\acquia_cms_site_studio\ExistingSite;
+namespace Drupal\Tests\acquia_cms_search\ExistingSite;
 
 use Drupal\Core\Entity\Query\QueryInterface;
 use Drupal\Tests\acquia_cms_common\ExistingSite\ContentTypeListTestBase;
 use Drupal\views\Entity\View;
 
 /**
- * Tests the "all places" listing page.
+ * Tests the "all events" listing page.
  *
  * @group acquia_cms
  * @group acquia_cms_site_studio
@@ -15,25 +15,25 @@ use Drupal\views\Entity\View;
  * @group pr
  * @group push
  */
-class PlaceListTest extends ContentTypeListTestBase {
+class EventListTest extends ContentTypeListTestBase {
 
   /**
    * {@inheritdoc}
    */
-  protected $nodeType = 'place';
+  protected $nodeType = 'event';
 
   /**
    * {@inheritdoc}
    */
   protected function getView() : View {
-    return View::load('places');
+    return View::load('events');
   }
 
   /**
    * {@inheritdoc}
    */
   protected function visitListPage($langcode = NULL) : void {
-    $page = $langcode ? "/$langcode/places" : "/places";
+    $page = $langcode ? "/$langcode/events" : "/events";
     $this->drupalGet($page);
   }
 
@@ -41,7 +41,7 @@ class PlaceListTest extends ContentTypeListTestBase {
    * {@inheritdoc}
    */
   protected function getQuery() : QueryInterface {
-    return parent::getQuery()->sort('title');
+    return parent::getQuery()->sort('field_event_start')->sort('title');
   }
 
 }
